Change to printout
Add header and algo printout

+++ Mini - MASSIVE/src/AnimationThread.java
@@ -1,8 +1,5 @@
 import javax.swing.*;
-import java.io.BufferedWriter;
-import java.io.FileWriter;
-import java.io.IOException;
-import java.io.PrintWriter;
+import java.io.*;
 import java.util.ArrayList;
 
 /**
@@ -61,24 +58,58 @@ public class AnimationThread extends Thread {
                                     otherSwarm.numAlive() + " " + otherSwarm.getArmyName() + " remain.\n" +
                                     "Algo " + otherSwarm.swarmAlgo + " beat Algo " + currSwarm.swarmAlgo + ".");
                     toggleAnimation();
-                    FileWriter fw = new FileWriter("src/results.txt", true);
-                    BufferedWriter bw = new BufferedWriter(fw);
-                    PrintWriter pw = new PrintWriter(bw);
-                    String resultsString = ticks + ", " + currSwarm.getArmyName() + ", " + otherSwarm.numAlive() + ", " + otherSwarm.swarmAlgo + ", " + currSwarm.swarmAlgo;
-                    //BufferedWriter writer = new BufferedWriter(new FileWriter("src/results.txt", true));
-                    pw.println(resultsString);
-                    pw.flush();
+                    File tempFile = new File("src/results.txt");
+                    if(!tempFile.exists()){
+                        String addInfo = "Ticks;  Losing Team; Number of Drones Left Alive; Winning Drone Num; Losing Drone Num";
+                        FileWriter fw = new FileWriter("src/results.txt", true);
+                        BufferedWriter bw = new BufferedWriter(fw);
+                        PrintWriter pw = new PrintWriter(bw);
+                        String resultsString = ticks + ", " + currSwarm.getArmyName() + ", " + otherSwarm.numAlive() + ", " + findName(otherSwarm.swarmAlgo) + ", " + findName(currSwarm.swarmAlgo);
+                        //BufferedWriter writer = new BufferedWriter(new FileWriter("src/results.txt", true));
+                        pw.println(addInfo);
+                        pw.println(resultsString);
+                        pw.flush();
 
-                    pw.close();
-                    bw.close();
-                    fw.close();
+                        pw.close();
+                        bw.close();
+                        fw.close();
+                    }else{
+                        FileWriter fw = new FileWriter("src/results.txt", true);
+                        BufferedWriter bw = new BufferedWriter(fw);
+                        PrintWriter pw = new PrintWriter(bw);
+                        String resultsString = ticks + ", " + currSwarm.getArmyName() + ", " + otherSwarm.numAlive() + ", " + findName(otherSwarm.swarmAlgo) + ", " + findName(currSwarm.swarmAlgo);
+                        //BufferedWriter writer = new BufferedWriter(new FileWriter("src/results.txt", true));
+                        pw.println(resultsString);
+                        pw.flush();
+
+                        pw.close();
+                        bw.close();
+                        fw.close();
+                    }
 
                     //writer.close();
                 }
             }
         }
     }
+    public static String findName(int num){
+        switch (num) {
+            case 0: //Flanking algorithm before initializing
+                return "Leeroy Jenkins";
+            case 1:
+            case 3:
+            case 4:
+                return "Flank";
+            case 2:
+                return "Fire Teams";
+            case 5:
+                return "Brain Swarm";
 
+            default:
+                //goto algo 0
+        }
+        return "Unknown Algo";
+    }
 
     // run() method is invoked automatically by the Java VM
     public void run() {
